Move menuLayoutsAvailable to WindowCapability
