+++ android/sdl_android/src/androidTest/java/com/smartdevicelink/test/rpc/datatypes/DisplayCapabilitiesTests.java
@@ -43,7 +43,6 @@ public class DisplayCapabilitiesTests extends TestCase{
         msg.setMediaClockFormats(Test.GENERAL_MEDIACLOCKFORMAT_LIST);
         msg.setScreenParams(Test.GENERAL_SCREENPARAMS);
         msg.setTemplatesAvailable(Test.GENERAL_STRING_LIST);
-        msg.setMenuLayoutsAvailable(Test.GENERAL_MENU_LAYOUT_LIST);
     }
 
     /**
@@ -60,8 +59,7 @@ public class DisplayCapabilitiesTests extends TestCase{
         List<MediaClockFormat> mediaClock = msg.getMediaClockFormats();
         List<TextField> textFields = msg.getTextFields();
         List<ImageField> imageFields = msg.getImageFields();
-        List<MenuLayout> menuLayouts = msg.getMenuLayoutsAvailable();
-        
+
         // Valid Tests
         assertEquals(Test.MATCH, Test.GENERAL_BOOLEAN, graphicSupported);
         assertEquals(Test.MATCH, Test.GENERAL_INT, numPresets);
@@ -72,8 +70,7 @@ public class DisplayCapabilitiesTests extends TestCase{
 		assertEquals(Test.MATCH, Test.GENERAL_MEDIACLOCKFORMAT_LIST.size(), mediaClock.size());
 		assertEquals(Test.MATCH, Test.GENERAL_TEXTFIELD_LIST.size(), textFields.size());
 		assertEquals(Test.MATCH, Test.GENERAL_IMAGEFIELD_LIST.size(), imageFields.size());
-		assertEquals(Test.MATCH, Test.GENERAL_MENU_LAYOUT_LIST.size(), menuLayouts.size());
-		
+
 		for(int i = 0; i < Test.GENERAL_STRING_LIST.size(); i++){
             assertEquals(Test.MATCH, Test.GENERAL_STRING_LIST.get(i), templatesAvailable.get(i));
         }
@@ -90,9 +87,6 @@ public class DisplayCapabilitiesTests extends TestCase{
             assertTrue(Test.TRUE, Validator.validateImageFields(Test.GENERAL_IMAGEFIELD_LIST.get(i), imageFields.get(i)));
         }
 
-        for(int i = 0; i < Test.GENERAL_MENU_LAYOUT_LIST.size(); i++){
-            assertEquals(Test.MATCH, Test.GENERAL_MENU_LAYOUT_LIST.get(i), menuLayouts.get(i));
-        }
         
         // Invalid/Null Tests
         DisplayCapabilities msg = new DisplayCapabilities();
@@ -107,7 +101,6 @@ public class DisplayCapabilitiesTests extends TestCase{
         assertNull(Test.NULL, msg.getScreenParams());
         assertNull(Test.NULL, msg.getTemplatesAvailable());
         assertNull(Test.NULL, msg.getTextFields());
-        assertNull(Test.NULL, msg.getMenuLayoutsAvailable());
     }
     
     public void testJson(){
@@ -123,7 +116,6 @@ public class DisplayCapabilitiesTests extends TestCase{
             reference.put(DisplayCapabilities.KEY_TEXT_FIELDS, Test.JSON_TEXTFIELDS);
             reference.put(DisplayCapabilities.KEY_IMAGE_FIELDS, Test.JSON_IMAGEFIELDS);
             reference.put(DisplayCapabilities.KEY_SCREEN_PARAMS, Test.JSON_SCREENPARAMS);
-            reference.put(DisplayCapabilities.KEY_MENU_LAYOUTS_AVAILABLE, JsonUtils.createJsonArray(Test.GENERAL_MENU_LAYOUT_LIST));
 
             JSONObject underTest = msg.serializeJSON();
             assertEquals(Test.MATCH, reference.length(), underTest.length());
@@ -163,7 +155,7 @@ public class DisplayCapabilitiesTests extends TestCase{
                 	Hashtable<String, Object> hashTest= JsonRPCMarshaller.deserializeJSONObject(underTestArray);
                     
                     assertTrue(Test.TRUE, Validator.validateScreenParams(new ScreenParams(hashReference), new ScreenParams(hashTest)));
-                } else if(key.equals(DisplayCapabilities.KEY_MEDIA_CLOCK_FORMATS) || key.equals(DisplayCapabilities.KEY_MENU_LAYOUTS_AVAILABLE)){
+                } else if(key.equals(DisplayCapabilities.KEY_MEDIA_CLOCK_FORMATS)){
                     JSONArray referenceArray = JsonUtils.readJsonArrayFromJsonObject(reference, key);
                     JSONArray underTestArray = JsonUtils.readJsonArrayFromJsonObject(underTest, key);
                     assertEquals(Test.MATCH, referenceArray.length(), underTestArray.length());
